Implement FailOperationEnqueuing and SpawnedTaskCount.
Related to LP 1953077. Similar to FailOperation, but doesn't set the operation status prematurely. If it is set here, the actions which did start will not be able to complete. Which causes the uniter to go into an infinite loop restarting on error. The spawned-task-count is updated as well to allow for the operation to complete.