Use FailOperationEnqueuing instead of FailOperation.
Related to LP 1953077. If not all actions enqueue, update the spawned-task-count and log the failures in the operation. Setting the operation status to fail, or an inaccurate spawed-task-count will cause the operation never to complete. Worse, unit agents will get stuck in a loop of uniter failures with state changing too quickly.

There are two types of enqueuing failures, one were an action doc is created and one where it isn't. The case where one isn't will be addressed here.

Also, capture all enqueuing failures, not just the first or last.