Adds LXD server certificate credential validation.
When Juju users call add-credential with their own LXD config.yaml they
need to make sure they make no mistakes with the paths to the various
pki components required.

It's possible for a user to set the wrong file for the server
certificate and for Juju to accept it and try and start using it for
secure communications with a LXD daemon.

What this change does is help to identify when mistakes like this have
been made by inspecting the LXD server certificate to make sure it
exists and that the certificate has been signed for server
authentication.

This is about the sum of all the checks we can perform on this
certificate without hard coding internal logic of LXD into Juju that
could change and introduce it's own further complications.