Add support for arch ppc64el.
For legacy reasons Juju uses arch ppc64el instead of ppc64le. This
change allows the makefile to accept both formats and continue working.
This is critical for Juju to function for older versions and to maintain
our CI systems.

In future versions of Juju we can look at dropping this support but it
would require a major version change.