Update output on individual action enqueuing error.
As actions are sent in bulk, erroring on a single enqueuing error
prevents the user from getting data on any other errors and/or the
successfully enqueued actions. Especially difficult with a quick return
as the user won't get the action ids to gather final status of
successfully enqueued actions.