Make parent worker not forcefully shut down child application worker
Trying to forcefully shutdown the application worker meant there was
a race condition when removing an application: if the parent worker
got the event first, it would shut down the app worker via
StopAndRemoveWorker (which would end up causing the app worker to take
the "case <-a.catacomb.Dying()" in its select and exit before it'd
had a chance to call a.dead which would ClearApplicationResources).

Just let the child app worker shut itself down gracefully -- it checks
for dead and handles that already.