Start child worker if app is dead so it goes through the dead() routine
Currently if for some reason the system gets into a funny state and
the child application worker dies before it's had a chance to go
through the a.dead() routine the application will never be deleted
from state.

Instead, only ignore if the application is not found (has already gone
away), but if it still exists and is dead, start the worker, which will
then go through the dying+dead sequence.