Direct controller-worker to controller API proof-of-concept
This adds a new controllerapi package (name not final) which is
intended for use in controller workers like caasapplicationprovisioner.
In this proof-of-concept I've only used it for the Life function, but
obviously other functions could be added in a similar way.

Note that all details, including package location and naming, are open
to debate. The important thing is the concept: being able to call
controller functions from controller workers simply, with these
advantages:

* No need to write client or server (facade) boilerplate functions.
* Function name and args are properly type checked by the Go compiler.
* No need for these controller-to-controller calls in the same process
  to be going over the network.

Also note that while this compiles, it doesn't work yet! I spent a
small amount of time debugging it, but didn't finish as I'm time-boxing
this PoC. I'm not sure why it's not working -- something's not getting
wired up correctly. But hopefully it shows the intent.