Fix logic in notAllocating (and rename it to avoid double negative)
First, rename the confusingly-named notAllocating function to isReady.
Then, fix the logic so it returns false when it should: specifically in
this case when the unit is allocating or in error before startup
(for example in the case of a container image not found).

See: https://bugs.launchpad.net/juju/+bug/1948695