Fixes a bug in for icmp rules in openstack.
When exposing applications in Openstack with icmp rules the Openstack
provider copies the port values from Juju internal rules to that of the
Openstack rules. However openstack expects port values to be nil when
making ICMP rules.

This change resolves this by not setting port values for ICMP proto's
and also by translating Openstack rules back to Juju rules by setting
the port to -1.

Fixes LP1970295