Fixes expose ipv6 issue in Openstack.
It's possible that when exposing an application in Juju that is deployed
across an Openstack provider the provider will only expost v4 cidr's and
not v6 cidr's. This happens because the cidr list to expose to is empty
so the provider defaults it off to v4 only.

This change doesn't reflect what happens elsewhere. When looking at the
facade for expose you can see that v6 cidr ranges are added to an expose
when none are provided.

This change now adds both v4 and v6 cidr ranges should none be provided
by the user.