Silence issue with teardown in kubeflow integration test
There was an error in teardown due to a bug in the
metacontroller-operator charm's "remove" hook:
https://github.com/canonical/metacontroller-operator/issues/5

For now, just avoid doing a proper teardown (testing that charm
"remove" hooks work is not what we're trying to test here) and let the
Microk8s teardown clean up.

I've also added a "juju status" in the failure case to show what
applications it was stuck on. (In this case I was able to repro
locally to find that it was the metacontroller-operator charm.)