Fixed broken AKS bootstrap.
AKS has changed the name of their provisioner used in new clusters.
Because of this Juju is unable to bootstrap to newly created AKS
clusters with the error message 'preferred storage "kubernetes.io/azure-disk" not available'.

What this change does is introduce a quick flag for AKS and the other
cloud clusters telling Juju that it is ok to use the default storage
provider. We have done the fix this way in 2.9 because of several bugs
found in the storage processing logic for Juju.

A proper and greater fix will land in 3.0 to address the short comings
of storage.

Fixes lp1976434