Add a juju_apiserver_build_info metric
As per guidance from Prometheus, we attach the version of the controller
onto a special metric. This always has a Gauge value of 1.0 and the
labels are used to indicate what version, etc we are running. This is
all known at compile time, so it can be all static info. We add all the
items that we have on the `running jujud` line for now.