Reworks the logic used for picking K8s storage.
What kicked off this change was the Juju was know longer able to
bootstrap to Azure clusters because the predefined notion Juju had about
what storage to use by default was know longer correct because Azure had
changed the naming of their provisioner's.

Upon closer inspection of the logic the code had a number of short
comings and bugs that did not allow it operate in the manner it was
originally intended to do.

This change is big but simplifies the logic used for finding storage in
Kubernetes now. Specifically cluster storage is now laid out
in provider preferences where the provider can find the best matching
storage class based on it's list. This allows the underlying cloud code
to be more agile and tolerate change from the underlying cloud and also
enshrines the logic used previously into a more robust readable
solution.