Check and branch on cloud type when remove-cloud
Previously remove-cloud would throw a generic error for clouds that
aren't personal (i.e. not in clouds.yaml) on the client.

Be a bit smarter and find all clouds listed, removing personal ones as
usual, but throwing a clearer error for built-in clouds

Also use gomock generated mocks in cloud/remove_test.go for the
RemoveCloudAPI. Also construct a mock provider that can return built-in
clouds to test the above.

Resolves https://bugs.launchpad.net/juju/+bug/1977871