Suppresses lost connect to pod error
Due to the way Kubernetes port forwarding works in the client api we are
forced to deal with klog errors that arise. A lot of these errors we
don't care about because our code retries gracefully. This change
introduces another error message to suppress from klog for Kubernetes
port forwarding.