Rename `juju relate` to `juju integrate`
The aliases are still available (add-relation, relate), but this makes
`juju integrate` the primary name for the command.
It also updates the help text of `juju integrate` along with the obvious
error messages.

There is a lot more to be done, but this is a first step towards how we
might explain documentation around integrations.