Add SetDownloadedIDAndHash and use.
The async charm downloader requires a path to add the charm origin ID
and hash to state.

The charm ID is needed after download of the charm and deployment as
most refresh actions on the refresh api only work with an ID of a charm,
rather than the charm name. Thus we do not save the charm name in the
charm origin. This will allow the charm author to change the charm name
in the future without impact to current deployments.