Fix cli/local_charms test due to deleted GitHub repo
These tests use https://github.com/lampkicking/charm-ntp, which stopped
working on the 14th of July (presumably it was deleted). So go back to
using https://git.launchpad.net/ntp-charm, which is the charm used on
Charmhub and presumably more stable. The test used to use this URL
until it was changed in
https://github.com/juju/juju/commit/c508cfad05d20a6f8b56d0768cae67ae22c865d8

Also add "-f" to "rm version" as the version file may or may not exist.

For some reason locally I get a different error than the test does,
presumably because of how I've configured git (with a username?) vs
how it's configured in the test infra. I get:

$ git clone --depth=1 --quiet https://github.com/lampkicking/charm-ntp.git
remote: Repository not found.
fatal: repository 'https://github.com/lampkicking/charm-ntp.git/' not found

But the test failure shows:

fatal: could not read Username for 'https://github.com': No such device or address

I'm pretty sure it's the same underlying cause, though. We'll soon see.