Do 3 retries when we get "EOF" error from Charmhub API
We get these weird errors from Charmhub every so often, for example:

ERROR resolving with preferred channel: Post "https://api.charmhub.io/v2/charms/refresh": EOF

These aren't valid HTTP responses, so we can't use the existing
juju/http retry logic as it only works for valid HTTP responses with
certain 40x and 50x status codes. This is an empty TCP response.

So retry 3 times (4 attempts) in this EOF case. We need to read in POST
bodies up-front because the http library will be reading/copying them
to the network more than once if a retry occurs. But they shouldn't be
huge, so this seems reasonable.

I used a test client and net.Listen server to reproduce this exact
case, to ensure the io.EOF test works on the real responses we got.
Here is the code for that:

* Client: https://pastebin.canonical.com/p/4RGNKggrK6/
* Server: https://pastebin.canonical.com/p/jtFfNP3GFF/