Clean up remaining doc.go issues with "generated GoMock package"
This removes the "Package X is a generated GoMock package." from the
package comments. For the most part I've done this by moving the mocks
to their own mocks/ directory. For the testing/ packages where the
mocks need to be in the package I've added -write_package_comment=false
to the mockgen command line.