Fix charm and bundle series selection differs.
LP 1983581. When deploying a bundle, there is not need to do series
selection twice for charmhub charms. Once in addcharm and again
independently in addapplication. Charmhub charms are downloaded and
saved with a specific series, where as charmstore charms were not. Use
the series selected at addcharm when deploying a bundle.

Includes a small fix for a regression introduced when selecting a series
when the Resolver did not. The series must be added to the url as well.