Fixes error messages when adding k8s models.
When adding k8s models where there already exists a namespace with the
same name we run into two scenarios. The first is for an existing
namespace that is owned by another Juju controller and the second is for
namespaces that have been created in a k8s clusters outside the control
of Juju.