juju info: get "supports" list from Charmhub instead of metadata.yaml
The series list from the metadata isn't what is available for deploy.
That comes from the Charmhub API response default-release bases (which
in turn comes from charmcraft and the manifest.yaml).

I also remove the ad-hoc isKubernetes function here and use
corecharm.IsKubernetes instead.

Unfortunately due to the existing TODO in convert.go:104, we don't get
the manifest.yaml and can't pass that to charmMeta / IsKubernetes, so
IsKubernetes returns the wrong value for container charms, e.g., the
likes of snappass-test. So snappass-test says it "supports: focal"
instead of what I'd expect, "supports: kubernetes". I think we'll need
to fix that in a follow-up change once ManifestYAML is provided by the
Charmhub API.

This commit also removes the unused Supports field from
bundleInfoOutput (it was never set so always omitted).