Resolve tests from failed merge attempt.
This file is significantly different between 2.9 and 3.0, the
merge went sideways. Reverted back to
6193bc6c961da225641bd5f664d868dc83a94481,
then added changes to be compatible with the changes made in
PR 14784, sha 75cad4d54ebc0cd5559524c52630cd6f93eb20fe.

Added TODOs to remove the updates for the tests to pass while upgrades
from 2.9 to 3.0 are not supported.