Update IAAS machine manifolds to remove unconverted-api-worker.
The unconverted-api-worker is being split into three pieces. machine-setup is short lived and does setup on the machine which required an API Connection. The other two are container provisioner workers for lxd and kvm.

The functionality of the container-watcher for setup is placed into the container provisioner worker. It waits for a container to be provisioned, configures the container system on the machine and starts an container provisioner worker.