Allows bootstrapping juju 3.0 to other arch's.
Due to a recent change in juju where the controller is now considered a
charm we were not passing along constraints to the application
deployment of the charm. This resulted in a case where the controller
charm was always trying to be deployed with amd64 instead of the arch
specified at Bootstrap time.

Now these arch constraints are passed along to the controller charm
deployment.