Fixes upgrade bugs for sidecar charms in 2.9
In the process of developing 2.9 we have made the mistake of placing pod
spec changes into the sidecar charms ensure method meant for newer
versions of Juju.

With this what ends up happening is that when a user upgrades a
controller the ensure method runs again for the sidecar application and
try's to start overwriting the pod spec with new addition's not
currently intended for the version of juju the model is running.

With this change we have placed these changes by gate's to check agent
versions when ensuring. We have also removed the upgrade path for
sidecard charms and moved the logic back to ensuring the state we expect
for different agent versions.

Fixes lp1997253