Tread through the context.Context
This is the start of threading through context.Context from the rpc
methods through to the actual method calling. Currently context.Context
is optional from the rpcreflect library, yet, it seems like a missed
opportunity to not use this. With the new database changes, we would
want to be able to cancel a sql transaction. Additionally, we could use
the context.Context for opentracing to help debug/diagnose running
issues.

The code makes a lot of changes, but it's all very mechanical.