[JUJU-2116]: Store default-base in model-config
Allow both a default-base and a default-series in model-config.
Default-series is deprecated, so we need to have both in the config at
once.

The next commit will reconcile both base and series in one location so
that we can get the correct information.

PreferredSeries is removed for PreferredBase

The following removes PerferredSeries so that PreferredBase is used.
This is quite a bit of work, as there are a lot of tests that expect to
work with series.

There is a few warts in this commit, but should be removed in the long
run. These are around a better distro info API, as that currently only
knows about series and not bases. Those should get resolved when we
remove more and more series related information from the APIs.

Add more tests

Ensure we only validate a base if it's set

driveby: remove comments

Attempt to keep both default-series and default-base insync

The following attempts to keep both default-series and default-base in
sync when performing model config changes. This should remove
inconsistent results.

This will obviously go away once series is dead.

Rebuild schema to export default-base

Add blank line after license

This is an interesting one, as I wrote the script to catch that.

driveby: fix the shellcheck

show_help exit 1 at the end, so it's not required to do it again, so
it's dead code - remove it.

Remove default-series altogether from model-config