JUJU-2351: Bootstrap to LXD vm
This set of changes allow bootstrapping to a LXD vm instead of a
container. Most of the work was already done, but it wasn't selecting
the right images to bootstrap from. Additionally, there are some issues
attempting to find the running status correctly for a VM. The simple
solution is to just implement a retry mechanism, with enough back off,
but there might be a better approach to this.

The tests don't currently pass, but future commits will fix those.

A simple QA to get up and running is the following:

```
$ juju bootstrap lxd --constraints virt-type=virtual-machine
```

We store a different alias for the VM image independently from a
container image, as they're different. LXD itself has an API for remote
lookup that doesn't require this to be explicit, but we do for the
alias. This means that we suffix all aliases with `/vm` to indicate the
difference. For containers, we leave as is, so that operators can reuse
the existing deployments without a new fetch.