Ensure that we set the virt-type in the lxd environ broker
When creating a new container type, we also want the virt-type to be
honoured. This does that by setting it in the environ broker and storing
it correctly for the instance data of the machine.

Additionally, to help with debugging, we also add the virt-type to the
formatHardware function, so it's obvious what is been changed.