Add virt-type when provisioning a machine
When provisioning a machine, the provisioner will set provisioned, which
will ensure that all the data for the machine is correctly added to the
instance data document. Sometimes this was missing when adding a
machine, fixing this then allows us to correctly see what the machine
was retrospectively in the db.