Error message with no current model selected
With the prior change of not creating a default model, there are now
occasions where the error messages are terrible from a UX perspective.
For example, bootstrapping then doing juju status is really, really
unhelpful. To better the UX, the solution is to just add a hint to use
juju switch or juju add-model.

The code is simple, it was just never done.