Preliminary panic fix for no primary nic vlan
I'm unsure if this is the whole fix when a primary nic VLAN is missing.
This just prevents a panic if the gomaasapi returns a set of device
interfaces that don't have a VLAN associated with it. This causes a
panic later on, as it's expected that a VLAN is always around.

The code now will just emit a warning stating that no subnet and no
primary NIC VLAN is available. I'm unsure what other options we have
here.