Normalise the virt-type when created container
We want to prevent any containers being created using the any type. The
any type is a empty string and lxd will choose which one it will pick.
To make the pick deterministic from the point of juju, we'll cast an
empty string to a container type. In the future we may want to model a
any type, but not atm.