Add comment about retrying on FilterContainers
The general gist of the comment is that we don't want to add the retry
logic at a higher level to know when to retry. Unfortunately we don't
have a way through the statuses to know if a virtual-machine container
is being created, but isn't quite alive. This uses a back-off to do
such.