Thread base through info and download
The following deprecates series in favour of base. Adds a warning if
you're using series. It removes all of the conversion between charmhub
base and series (which is a string) and core series base type. This
means the code is a lot simpler.

The output has no series in it and refers to base. All output is now
bases.

There is still some complexity around "all", but we will be able to drop
that in favour of an empty base type.