Parse series as bases for list images
Deprecate series for list images and then ensure that series are parsed
as bases. The new series of bases are then sent to the client. The base
is no longer validated for the version, this is important for subsequent
changes in the future, where we need to be able to send custom images
that aren't locked to a prescribed series version.