Remove LXD server checking when applying constraints
Removal of LXD server checking for LXDs that are less than a given
server version isn't required. We're explicitly calling out that Juju
only works with LXD 5.0 or greater. Having code that checks for server
versions for older versions of LXD isn't ideal for keeping the code path
clean.

Additionally, logging out that the virt-type parsing failing with a log.