Exclude devices which don't have an address
Assigning link layer devices with no addresses to spaces fills the
default alpha space with lots of devices that aren't used. This means
that logic working on spaces expect the devices to be fully functional.
This isn't always the case. Preventing this then helps clean up the
alpha space, along with removing unwanted devices, which can cause
panics later on.