Removes the deadline typed error
The deadline typed error for the lases manager was for the older
implementation of leases. This typed error was used to indicate when an
API didn't respond in a timely fashion.

Just removing the error is enough to ensure that we've not got any code
paths in the lease manager that aren't exercised.