Standardise jujud building
The makefile has a standardised way of building binaries and it's shame
that when dqlite was added (by me), we didn't use the new hotness. This
finally fixes that and we're almost the same with a tiny few exceptions.

Ideally we'll move to musl to build for different architectures, but
that's a bit more work. This is the first step (read as foundation),
before moving to zig.