Allow overriding shell machine args
The following allows the introduction of arguments to the Makefile
targets. The first one allows overriding the machine type (x86_64),
which is what we use from linux and the goarch is used for specifying
the pkg type from the go side.

This means that we don't have to introduce the `uname -m` concept to the
juju makefile side.