Allow consuming of related changes
The following changes make attaching a filesysystem more reliable after
a machine restart.

When a machine is restarted the order of changes from the watchers are
not guaranteed to be executed. Such that the go select case semantics
actually forces a shuffle if there a multiple select cases with buffered
channel values. This exacerbated when the machine is restarted and
relaunching the agent will get all the initial values from the all the
watchers. As this is random, it wasn't always obvious when this might
happen. It depends on the number of select case statements with values
and only if filesystem changes happens after filesystem attachment
changes. So if a filsystem changes is in a different order, but is
before filesystem attachment changes then you wouldn't see the
filesystem attachment issue.

The fix here is to consume any relational dependant changes before
working on child changes. As filesystem changes can be processed all at
once before processing the round of attachments, the change should
ensure an normalised ordering.