Return no value after waiting a short amount of time.
In an ideal scenario we wouldn't/shouldn't have to wait, but because you
might not get that information in the same iteration of the loop, we
should wait for a small amount of time before advancing. The correct
solution to this, is to ensure that when adding a file system
attachment, query the current state of the world ensuring everything is
in the correct state.

This isn't ideal, but better than having no information.