Bundle deploys respect model-constraints
When deploying a bundle with no constraints with in the application, we
should check the model constraints before attempting to fallback to the
default architecture.

This was originally coded like this because the aim was just to get the
code working. We never went back to make sure that the bundle correctly
checked every level of constraints in the right order. Those being:

 - user provided constraints
 - application constraints
 - model constraints
 - default constraints (architecture)

The code is really simple, just pass in the existing model constraints
from the model when working out what the application constraint should
be.