Allow build script to build ppc64le
The following fixes the build script to take into account building on a
powerpc64 little endian architecture. Originally I had to change a lot
to get this to work, but later found out we can force long-double
support for 64 architecture, reducing complexity.

Also the asm location on a powerpc machine is the long name and not the
short name, so we have to hack around that one as well.