Allow cross compiling jujud per GOARCH
The following allows cross compiling jujud via musl-cross-make.
musl-cross-make builds musl first for the correct architecture along
with it's dependencies:

  - binutils
  - gcc
  - gmp
  - mpc
  - mpfr
  - isl

The way the make file works is that if you're attempting to install
jujud locally, it will use musl for the architecture you have locally.
For building jujud it will use the musl-cross-make. It will build the
musl in $GOPATH/build/musl-$ARCH.

I've successfully built the following:

  - amd64
  - arm64
  - s390x
  - ppc64le

Unfortunately, I'm struggling to get riscv64 building, but I'd like to
get this landed first.