Build the musl cross library
The following builds the musl cross library ensuring that we can
correctly build jujud. There is a lot going on here to ensure we play
nicely.

Additionally, we need to cache the musl build for the os and arch, as
it's very expensive to do it every time we attempt to build jujud.