Remove Main test go file
We can no longer run jujud test suite like this, as we need to pass in
all the C libraries to build for this test. The better approach is to
move these sorts of tests to the bash integration tests.

I've made a bug[1] to track this change at a later date.

  1. https://bugs.launchpad.net/juju/+bug/2004182