Remove all the featuretests
The featuretests are tests which acted like integration tests, with the
aim of testing the piece of code end to end, but never did it in
isolation. The setup of a piece was done from not the top most layer,
but from a couple of layers down. It then modified the database to be in
a state that was either partially constructed or setup in
tautologically.

Additionally, the tests are located next to the actual pieces of code
they're trying to test. This means that some code is exposed just to be
able to either setup the test or test a given feature.

I think we can do better than this. I've made a bug to track this[1]

 1. https://bugs.launchpad.net/juju/+bug/2004278