Build jujud as a snap
The following builds jujud as a snap using musl. This follows some of
the build setup that we have in the build scripts, but we're dodging a
few builds by downloading the dependencies directly from s3 rather than
building them from scratch.

We can build all these things and in the long run we probably should,
but we need to just get this off the ground.

Additionally, we end up with all the dqlite dependencies in the snap as
well which we really don't want to do, so I'm unsure how to cleanly
remove them.