Allow testing on a mac
The following allows testing without dqlite on a platform that isn't
linux. This is the first step to allow the github actions to pass.

If you attempt to run a test on a package that requires dqlite then it
will fail, but there isn't much I can do with that atm. At least until I
can come with an alternative solution.