Install musl
Install musl to the _deps package for all musl builds. Prior to this we
installed musl in different places depending on the arch. This fixes
that so it's more easier to reason about. Everything should install to
_deps.

We're currently not building musl with -static flag, although that could
be made to happen, which would help keep the local system cleaner.