Saved by unit tests, ensure we do not panic on nil pointer.
Unit tests and SetCharm called by the Update application facade method will
not contain a charm origin. Ensure we don't panic on this in the real world.

The Update facade method is only kept for compatibility with older juju 2.x
clients. With juju 3.0, we can revaluate checking for an empty charm origin
in the facade and updating the unit tests.