Enable content addressable downloads
The following changes the file names to be content addressable. This
should help address some of the warts of the chicken and egg problem.
The problem is how do you distribute your new dqlite changes to others
without overriding the latest stable tar files. You can use a date, but
that prevents adding multiple libs for the same day, we could use a
timestamp, but doesn't really serve any more information than what we
already have on the filesystem. Instead if we use the sha256 of the file
itself as the file name, we can then have multiple ones. Additionally,
we may get the possibility of reducing how much we store if the files
built with the same dependencies build to the same sha (I doubt it
because of date/time impacting the builds).

The workflow for updating a lib then becomes:

 1. You build locally testing the changes.
 2. You push your sha256 file (tar.bz2 suffix) to s3
 3. Others can override the sha256 values in
 ./scripts/dqlite/scripts/dqlite/dqlite-install.sh
 4. Once verified those sha256 can be committed.

If the files aren't good, we can just remove them manually.

If we're worried about the s3 bucket becoming large, we can add folders
for released/proposed/develop etc.

As a driveby I've finally removed the one shell, so the make file should
be fully portable.