ChangeLog DDL
The following adds the change log DDL schema. The changelog represents a
way to collate the flow of changes from a set of table triggers to a
stream of data changes. These data changes represented in the change log
can that be queried at a later stage to know what's changed. This
information will be utilised to drive an event stream for watchers to
know what's changed.

This will be the sub system for driving the new juju watchers. This
schema was derived from the nu-juju-watchers[1] prototype.

The difference between this change and the prototype, is that we're
constraining values to give better type information and indexing
opportunities.

Additionally, I've fixed --build-agent when bootstrapping.

  1. https://github.com/SimonRichardson/nu-juju-watchers/blob/main/schema.go