Use INT as primary key for change_log
We need to use a INT as the primary key for the change log, one that
auto increments. This is so we can use it as a way of searching for
changes after a prior primary key. Otherwise we need to use the time and
that has issues when changing between controllers. The controllers time
could potentially have clock drift. In this case, we can just rely on
clocks as a potential way to sort and not as an offset selector.