Forces Juju to create IMDSv2 instances only.
Juju currently deploys machines into AWS EC2 with instance metadata that
works for both v2 and v1 of IMDS. V1 of IMDS is long fazed out of AWS
and is kept around for legacy purposes and provides an ongoing security
risk that Juju users won't tolerate.

With this change we strictly force new instances created by Juju to only
work with IMDSv2.

Writing regression tests for this in the current provider model is quite
difficult. This commit asks the user to validate the change manually.

As part of this PR I have checked cloud-init supports IMDSv2 for all of
the Ubuntu release we care about in 3.1.

Bug reference: https://bugs.launchpad.net/juju/+bug/1960568