Change stream worker
The change stream worker defines a worker that sits upon the db accessor
worker. It is expected that any changes pushed into the change log will
be picked up when the change stream loops around. The change stream
groups similar changes and coalesces the changes together. If a change
has the same UUID and the same type namespace and edit type then any
changes that arrive together will reference the last known one.

To optimise the query even more, we can store the last know valid id and
only querying changes after that id. So when the worker starts, it will
perform a scan of all the change log to gather that information. It
might be wise to actually do a first scan of the table on first
initialisation and only then dispatch changes there after. This can be
discussed in the PR.

The worker hands out a change stream per namespace. Currently we only
have a controller database, we may want to see at a later stage if we
want to consolidate to just one stream for multiple namespaces
(databases).

When the change stream executes, it does this at a given polling tick.
That's currently set to 100ms, we ideally do not want to thrash the
database too much, but we might want to look into adaptive polling
techniques. If we have seen a lot of data, keep polling at a smaller
interval, before gradually moving back out to a larger polling interval.
I've not implemented that as yet, but left a TODO for later.

This is only one of the foundational parts of the dqlite watcher setup:

  - db accessor (dqlite)*
  - change stream*
  - event queue
  - watchers

Currently, the db accessors and the change stream is done. The event
queue will follow on from that.