Remove unexposed API methods
The following removes unexposed API methods that were added to hide
implementation details about what methods are exposed and which are not.

By definition they're empty and do not conform to the Juju API. The jsonschema
package correctly handles the case when a struct{}{} is used in an API
definition, but we then take that information verbatim.

The following just removes any empty definitions.