Remove methods not supported
The following removes methods from the object type that the API server
doesn't want to advertise. These methods are normally of the following
type:

    func(Facade) DoSomething(_ struct{}{}) {}

They normally to deprecate or hide a method on the facade so you can no
longer use them. Originally this was only for agents, but with the new
aggressive (considered good) changes to client facades, this pattern
needs to not leak into the schema. We should stop advertising the method
completely in the json, so other clients (pylibjuju) can no longer use
them.