Build jujud on a mac for linux
The following allows the building of jujud on a mac for linux. This uses cross
compiling to build jujud. At no point is jujud going to support building on
darwin directly, so to get this to work you can use GOOS=linux and the GOARCH
of either amd64 or arm64. It won't support s390x or ppc64le as that's a lot of
work for very little gain. If someone wants to do that, you'd have to do it in
a VM to get the tool chain setup.

There are a lot of tweaks to the musl installation, but it should be gated by
the is_darwin function.

By default the implementation requires the brew cross musl[1] package. It will
install both amd64 and arm64, which will take a few minutes longer to install,
but it means it's a lot easier to do it correct upfront, rather than try and
detect which one is installed.

To use --build-agent, we may have to push the GOOS and GOARCH through the
code path, or it might just work doing:

    GOOS=linux GOARCH=amd64 juju bootstrap lxd test --build-agent

It will require more investigation, which can be done in a follow up PR.

  1. https://github.com/FiloSottile/homebrew-musl-cross