Copy required methods from lxc/lxd shared package
The following copies the required methods that juju needs from the
shared package. We need to do this, because there are some CGO enabled
methods that are now being triggered when we build jujud with
CGO_ENABLED=1.

Ideally we would change upstream to move the CGO to a sub package, but
that dictates how Juju uses lxc/lxd package as a client, which isn't
ideal.

Using these following changes should then allow us to migrate some of
the header issues that we currently have when building jujud.

I've put the changes under a dual license, as Apache 2.0 should be
compatible with AGPLv3, but IANAL.

Lastly, as we're copying over the generate mem cert, we should pay
attention to this original source when upgrading the lxc/lxd package in
go.mod, as we might find that we're no longer compatible.