FileNotifyWatcher watch file changes
The following is a file notify watcher, that wraps the inotify syscall
on linux. This is a much more efficient implementation that the wrench
package, as we're leaning onto the kernel to emit an change if there is
any change to a folder.

We will watch a given folder for file creation/deletion and then perform
a block/unblock on the change stream.

The premise behind this is to be able to control the changestream
without using the database. This is currently just a open/close valve,
but I don't see why we can't use it for throttling the changestream to
certain events or amount of events.

There is a lot of plumbing because of the basic boiler plate of workers,
but I do believe this should be resilient enough for what we want.